feat: Add some custom exceptions for errors related to Rachas
Many parts of the validation code have been moved out of the controller to the RachaService for consistency. All tests passed

+++ src/main/java/ifsul/agileproject/rachadinha/service/impl/RachaServiceImpl.java
@@ -11,6 +11,8 @@ import ifsul.agileproject.rachadinha.domain.dto.RachaUpdateDTO;
 import ifsul.agileproject.rachadinha.domain.entity.Racha;
 import ifsul.agileproject.rachadinha.domain.entity.Status;
 import ifsul.agileproject.rachadinha.domain.entity.User;
+import ifsul.agileproject.rachadinha.exceptions.ForbiddenUserException;
+import ifsul.agileproject.rachadinha.exceptions.RachaNotFoundException;
 import ifsul.agileproject.rachadinha.mapper.RachaMapper;
 import ifsul.agileproject.rachadinha.repository.RachaRepository;
 import ifsul.agileproject.rachadinha.service.RachaService;
@@ -23,7 +25,7 @@ public class RachaServiceImpl implements RachaService {
 
   @Autowired
   private RachaMapper rachaMapper;
-
+  
   @Override
   public Racha saveRacha(RachaRegisterDTO rachaDTO) {
     Racha racha = rachaMapper.apply(rachaDTO);
@@ -37,6 +39,9 @@ public class RachaServiceImpl implements RachaService {
 
   @Override
   public Optional<Racha> findRachaById(Long id) {
+    if (!rachaRepository.existsById(id)) {
+      throw new RachaNotFoundException(id);
+    }
     return rachaRepository.findById(id);
   }
 
@@ -46,6 +51,21 @@ public class RachaServiceImpl implements RachaService {
 
   @Override
   public void deleteRachaById(Long id) {
+    if (!rachaRepository.existsById(id)) {
+      throw new RachaNotFoundException(id);
+    }
+    rachaRepository.deleteById(id);
+  }
+
+  @Override
+  public void deleteRachaById(Long id, Long loggedUserId) {
+    if (!rachaRepository.existsById(id)) {
+      throw new RachaNotFoundException(id);
+    }
+    Racha racha = rachaRepository.findById(id).get();
+    if (racha.getOwner().getId() != loggedUserId) {
+      throw new ForbiddenUserException(loggedUserId);
+    }
     rachaRepository.deleteById(id);
   }
 
@@ -67,12 +87,27 @@ public class RachaServiceImpl implements RachaService {
 	}
 
   @Override
+  public Racha updateRacha(RachaUpdateDTO rachaUpdateDTO, Racha racha, Long loggedUserId) {
+    if (racha.getOwner().getId() != loggedUserId) {
+      throw new ForbiddenUserException(loggedUserId);
+    }
+    return updateRacha(rachaUpdateDTO, racha);
+  }
+
+  @Override
   public Racha findRachaByStatus(Status status) {
     return rachaRepository.findByStatus(status);
   }
 
   @Override
   public List<Racha> findRachaByOwner(User owner) {
+    return rachaRepository.findByOwner(owner);
+  }
+
+  @Override
+  public List<Racha> findRachaByOwner(Long id) {
+    User owner = new User();
+    owner.setId(id);
     return rachaRepository.findByOwner(owner);
   }
 }
